refactor(board): modify board UI layout
